requirements-backup.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.12
aiosignal==1.3.2
alabaster==0.7.16
alembic==1.16.2
amqp==5.3.1
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.9.1
astroid==3.0.3
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncping3==3.1.2
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.0.4
Babel==2.14.0
backcall==0.2.0
backports.tarfile==1.1.0
beautifulsoup4==4.12.3
billiard==4.2.1
binaryornot==0.4.4
bitarray==2.9.2
bitstring==4.1.4
black==24.4.0
bleach==6.1.0
bleak==0.21.1
blessed==1.20.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
celery==5.5.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpickle==3.0.0
cmudict==1.0.33
colorama==0.4.4
coloredlogs==15.0.1
colorlog==6.6.0
comm==0.2.2
cookiecutter==2.6.0
cpyrt==0.1.0
cryptography==42.0.5
cssselect==1.3.0
Cython==3.0.0
dbus-fast==2.21.1
dbus-python==1.2.18
debugpy==1.8.1
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
devgoldyutils==3.0.0
diff-match-patch==20230430
dill==0.3.8
distro==1.7.0
Django==5.2.4
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
docstring-to-markdown==0.15
docutils==0.21.1
editor==1.6.6
email_validator==2.2.0
enum-compat==0.0.3
evdev==1.6.1
exceptiongroup==1.2.1
executing==2.0.1
fasteners==0.14.1
fastjsonschema==2.19.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
flake8==6.1.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flet==0.19.0
flet-core==0.19.0
flet-runtime==0.19.0
fqdn==1.5.1
freetype-py==2.4.0
frozenlist==1.7.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.14.0
hsluv==5.0.4
httpcore==0.17.3
httplib2==0.20.2
httpx==0.24.1
humanfriendly==10.0
idna==3.7
imagesize==1.4.1
img-to-pdf==1.0.2
img2pdf==0.5.1
importlib_metadata==7.1.0
importlib_resources==6.5.2
inflection==0.5.1
iniconfig==2.0.0
inquirer==3.3.0
intervaltree==3.1.0
ipykernel==6.29.4
ipython==8.23.0
isoduration==20.11.0
isort==5.13.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jedi==0.17.2
jeepney==0.8.0
jellyfish==1.0.3
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
keyring==25.1.0
kombu==5.5.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.3
lxml==5.2.2
macaroonbakery==1.3.1
Mako==1.3.10
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
meson==1.4.1
mistune==3.0.2
mne==1.6.1
mne-lsl==1.1.1
monotonic==1.6
more-itertools==10.2.0
MouseInfo==0.1.3
mov-cli-test==1.1.7
multidict==6.4.4
muselsl==2.2.2
MuseLSL2 @ https://github.com/DominiqueMakowski/MuseLSL2/zipball/main
mutagen==1.47.0
mypy==1.16.1
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
newspaper3k==0.2.8
nltk==3.9.1
notebook==7.0.3
notebook_shim==0.2.3
numpydoc==1.7.0
oauthlib==3.2.2
outcome==1.3.0.post0
overrides==7.4.0
packaging==24.0
pandas==2.0.3
pandocfilters==1.5.1
parso==0.7.0
pathspec==0.12.1
pdftext==0.6.2
pexpect==4.9.0
pickleshare==0.7.5
pikepdf==9.7.0
pillow==10.2.0
pillow_heif==0.15.0
platformdirs==4.2.0
pluggy==1.5.0
pooch==1.8.0
pop-transition==1.1.2
prometheus-client==0.17.1
prompt-toolkit==3.0.43
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycairo==1.20.1
pycodestyle==2.11.1
pycparser==2.22
pycryptodomex==3.20.0
pycups==2.0.1
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydbus==0.6.0
pydocstyle==6.3.0
pyflakes==3.1.0
pygame==2.5.2
pygatt==4.0.5
PyGetWindow==0.0.9
Pygments==2.17.2
PyGObject==3.42.1
PyJWT==2.3.0
pylint==3.0.4
pylint-venv==3.0.3
pyls-black==0.4.7
pyls-spyder==0.4.0
pylsl==1.16.2
pymacaroons==0.13.0
PyMsgBox==1.0.9
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.0
pyperclip==1.8.2
pyphen==0.17.2
pypng==0.20220715.0
PyQt5==5.12.3
PyQt5-Qt5==5.15.2
PyQt5-sip==12.13.0
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0
pyqtgraph==0.13.3
PyQtWebEngine==5.12.1
PyQtWebEngine-Qt5==5.15.2
PyRect==0.2.0
pyRFC3339==1.1
PyScreeze==0.1.30
pyserial==3.5
PySocks==1.7.1
pytest==8.1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-decouple==3.8
python-dotenv==1.1.0
python-gnupg==0.4.8
python-json-logger==2.0.7
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
python-libdiscid==2.0.1
python-lsp-black==2.0.0
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.11.0
python-slugify==8.0.4
python-xlib==0.29
python3-xlib==0.15
pytoolconfig==1.3.1
pytubefix==5.6.3
pytweening==1.0.7
pytz==2022.1
pyuca==1.2
pyxdg==0.28
PyYAML==6.0.1
pyzmq==26.0.2
QDarkStyle==3.2.3
qrcode==7.4.2
qstylizer==0.2.3
QtAwesome==1.3.1
qtconsole==5.5.1
QtPy==2.4.1
rapidfuzz==3.9.3
readchar==4.1.0
redis==6.2.0
referencing==0.34.0
regex==2024.11.6
repath==0.9.0
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
requests-file==2.1.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rgain3==1.1.1
rich==13.7.1
rope==1.13.0
rpds-py==0.18.0
rsa==4.9.1
Rtree==1.2.0
ruff==0.12.2
runs==1.2.2
scikit-learn==1.4.0
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.3
selenium==4.31.0
Send2Trash==1.8.2
sessioninstaller==0.0.0
sgmllib3k==1.0.0
shellingham==1.5.4
Shredder===2.9.0.Odd.Olm
six==1.16.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.5
Sphinx==7.3.7
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
spyder==5.5.0
spyder-kernels==1.10.3
spyder-notebook==0.5.1
spyder-terminal==1.2.2
SQLAlchemy==2.0.41
sqlparse==0.5.3
ssh-import-id==5.11
stack-data==0.6.3
superqt==0.6.3
systemd-python==234
tabulate==0.9.0
terminado==0.17.1
text-unidecode==1.3
textdistance==4.6.1
textstat==0.7.7
thefuzz==0.22.1
threadpoolctl==3.2.0
three-merge==0.1.1
tinycss2==1.2.1
tinysegmenter==0.3
tldextract==5.3.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
tornado==6.4
tqdm==4.62.3
traitlets==5.14.3
trio==0.29.0
trio-websocket==0.12.2
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ujson==5.9.0
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.1
vine==5.1.0
vispy==0.14.1
wadllib==1.3.6
watchdog==4.0.0
wcwidth==0.2.13
webcolors==1.13
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.1
wget==3.2
whatthepatch==1.0.5
whitenoise==6.9.0
Wikipedia-API==0.8.1
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
wurlitzer==3.0.3
xdg==5
xkit==0.0.0
xmod==1.8.1
yapf==0.40.2
yarl==1.20.1
yt-dlp==2025.3.31
zipp==3.18.1
